Widening matching function width to drop number of NCCs in single cheby multitrope.
This drops the NCCs from ~200 to ~100 for the n_rho_cz=1, n_rho_rz=1 cases, with big impacts on memory per core requirements and on time for steps that require LU factorization (changes to dt), where we go from ~3 min to ~1m:20s for a nz=1024 case.  The constant dt steps appear to be almost unaffected.